requirements.txt
Flask==0.10.1
Jinja2==2.7.3
Landscape-Client==14.12
MarkupSafe==0.23
PAM==0.4.2
Pillow==2.8.1
Twisted-Core==14.0.2
Werkzeug==0.10.4
argparse==1.2.1
characteristic==0.1.0
chardet==2.2.1
colorama==0.3.1
configobj==4.7.2
html5lib==0.999
idna==0.9
iotop==0.6
itsdangerous==0.24
pyOpenSSL==0.13.1
pyasn1==0.1.7
pyasn1-modules==0.0.5
pyserial==2.6
python-apt==0.9.3.10ubuntu1
python-debian==0.1.22
requests==2.3.0
service-identity==1.0.0
six==1.7.3
ssh-import-id==3.21
urllib3==1.8.3
virtualenv==12.1.1
wheel==0.24.0
wsgiref==0.1.2
zope.interface==4.1.1
